requirements.txt
asgiref==3.2.7
astroid==2.4.1
bleach==3.3.0
certifi==2020.4.5.1
cffi==1.14.5
chardet==3.0.4
colorama==0.4.3
cryptography==3.4.7
defusedxml==0.7.1
dj-database-url==0.5.0
Django==3.0.6
django-allauth==0.44.0
django-appconf==1.0.4
django-countries==7.1
django-crispy-forms==1.9.0
django-daguerre==3.0.0
django-heroku==0.3.1
django-imagefit==0.6.2
django-widget-tweaks==1.4.8
docutils==0.17.1
gunicorn==20.0.4
idna==2.9
importlib-metadata==4.0.1
invoke==1.5.0
isort==4.3.21
keyring==23.0.1
lazy-object-proxy==1.4.3
mccabe==0.6.1
oauthlib==3.1.1
packaging==20.9
Pillow==7.1.2
pkginfo==1.7.0
psycopg2==2.8.5
pycparser==2.20
Pygments==2.9.0
PyJWT==2.1.0
pylint==2.5.2
pyparsing==2.4.7
python3-openid==3.2.0
pytz==2020.1
pywin32-ctypes==0.2.0
readme-renderer==29.0
requests==2.23.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
rfc3986==1.5.0
six==1.14.0
sqlparse==0.3.1
toml==0.10.0
tqdm==4.60.0
twine==3.4.1
urllib3==1.25.9
webencodings==0.5.1
whitenoise==5.0.1
wrapt==1.12.1
zest.releaser==6.22.1
zipp==3.4.1